pyproject.toml
[tool.poetry]
name = "equalify"
version = "0.1.0"
description = ""
authors = ["Sam Altman <sam@openai.com>"]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.38.0"
bcrypt = "^4.2.0"
beautifulsoup4 = "^4.12.3"
tqdm = "^4.66.5"
pymongo = {extras = ["srv"], version = "^4.9.1"}
python-dotenv = "^1.0.1"
openai = "^1.50.2"
scikit-learn = "^1.5.2"
seaborn = "^0.13.2"
plotly = "^5.24.1"
Faker = "^30.0.0"
wordcloud = "^1.9.3"
nltk = "^3.9.1"
gensim = "^4.3.3"
textblob = "^0.18.0"
textstat = "^0.7.4"
umap-learn = "^0.5.6"
hdbscan = "^0.8.38"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
